Remove /tmp/vconf-init mechanism
The vconf code has a left over, racey and broken, workaround to
have all vconf set operations attempt to wait for a file called
/tmp/vconf-initialized.  If a set for an in-memory key was
attempted when the file did not exist, then after nearly two
seconds of trying the end result is the backup key file was
created, but containing no data.

Because of this, and because vconf initialization does not happen
at image creation time, then all images are created with many
in-memory backup items containing empty files.  This results in
many subtle hard to debug bugs throughout the stack.

Change-Id: I3f9a4e1aa0f8267ccc07631c0eae2e2315ab2d17
Signed-off-by: Rusty Lynch <rusty.lynch@intel.com>